# Adds a binary IonReader implementation capable of incremental reads.