… src/com/amazon/ion/IonBufferConfiguration.java
… src/com/amazon/ion/IonReader.java
… src/com/amazon/ion/impl/IonReaderBinaryIncremental.java
… src/com/amazon/ion/impl/IonTypeID.java
… src/com/amazon/ion/impl/ReaderLookaheadBuffer.java
… src/com/amazon/ion/impl/ReaderLookaheadBufferBase.java
… src/com/amazon/ion/impl/ResizingPipedInputStream.java
… src/com/amazon/ion/impl/SymbolTokenImpl.java
… src/com/amazon/ion/impl/_Private_IonReaderFactory.java
… src/com/amazon/ion/impl/_Private_RecyclingStack.java
… src/com/amazon/ion/impl/_Private_Utils.java
… src/com/amazon/ion/impl/bin/IonRawBinaryWriter.java
… src/com/amazon/ion/impl/lite/IonDatagramLite.java
… src/com/amazon/ion/system/IonReaderBuilder.java
… test/AllTests.java
… test/com/amazon/ion/BinaryIonReaderIteratorSystemProcessingTest.java
… test/com/amazon/ion/BinaryReaderSystemProcessingTest.java
… test/com/amazon/ion/BinaryReaderWrappedValueLengthTest.java
… test/com/amazon/ion/BinaryTest.java
… test/com/amazon/ion/EquivsTestCase.java
… test/com/amazon/ion/GoodIonTest.java
… test/com/amazon/ion/IonTestCase.java
… test/com/amazon/ion/LoadBinaryIonReaderSystemProcessingTest.java
… test/com/amazon/ion/NopPaddingTest.java
… test/com/amazon/ion/RawValueSpanReaderBasicTest.java
… test/com/amazon/ion/ReaderChecker.java
… test/com/amazon/ion/ReaderMaker.java
… test/com/amazon/ion/ReaderSystemProcessingTestCase.java
… test/com/amazon/ion/RoundTripTest.java
… test/com/amazon/ion/SystemProcessingTestCase.java