@tgregg 2021-05-07
> There's a lot of complexity around saving and testing state that aims to avoid ever having to re-process an input byte if `moreDataRequired()` returns `true`. For example, if the buffered data ends in the middle of an annotation ID, the lookahead buffer can resume parsing the rest of the annotation ID on the next call.
> 
> I would wager that for most typical use cases the "insufficient data" case is comparatively rare, and that most calls to `next()` will be satisfied by the buffer. I suspect that a lookahead buffer implementation that was willing to re-parse bytes in the (uncommon) insufficient data case would not only be much simpler, it would probably perform better in general because it wouldn't need to save/test state for each byte.
> 
> Does saving/testing state byte-by-byte buy us other functionality that I've glossed over?

This is an interesting suggestion that I've now had time to think about for a while. I agree that "insufficient data" happens much less frequently than "sufficient data" when calling `moreDataRequired()`.

However, I wasn't able to find much complexity in the code that I'd be able to eliminate by re-parsing bytes after previously returning `true` from `moreDataRequired()`. Here is some of my thought process.

1. It doesn't apply to `readTypeId()`, which only reads a single byte. Either the byte is there or it isn't; there's nothing to re-parse.
2. It does apply to `readVarUInt()`, which is the only other place where bytes are actually parsed. At first, I was thinking simplifying this might allow us to get rid of the `VarUInt` class and `inProgressVarUInt` variable. However, we would then have the problem that we still need to return multiple pieces of information from the `readVarUInt()` method, namely:
** Is the VarUInt complete? If not, `moreDataRequired()` must return `true`.
** If the VarUInt is complete, what is its value *and* how many bytes composed the value? (The latter is needed when reading annotation wrappers.)
Therefore, I conclude that a reusable object like VarUInt is still the best way to convey this information, and keeping the in-progress value doesn't cost us anything. In fact, it is actually necessary in the case where the current value is being skipped due to exceeding the maximum size, because in that case bytes are read directly from the underlying InputStream and are never buffered. The value could not be re-parsed in that case because the bytes that composed it are gone.
3. At the extreme, if each call to `moreDataRequired()` were completely independent (i.e. if a previous call had to return `true`, the next call would have to re-read every byte from the value starting from the type ID), would that allow us to simplify the state machine? And I came to the conclusion that I don't think it would. We'd still have a loop where we progress through states, with certain states sometimes being skipped based on the outcome of previous states (e.g. if the value is unannotated). We'd also have the same problem as in 2) above, where values being skipped are never buffered.

I'm happy to continue discussing this if you can see something I've missed, though. If we can find a technique that significantly simplifies the code and provides better performance for the `moreDataRequired() -> false` case at the expense of the `moreDataRequired() -> true` case, I'm all for it.